Configure AWS S3 upload the image
Pitfalls 3
(1) Understand how to create users and set permissions in AWS. How to modify these keys on the command line using vim, `EDITOR=vim bin/rails credentials:edit`
(2) When configuring storage.yml, you need to set the time zone endpoint, set the storage bucket name, and finish the configuration.
(3) The default upload is not an image, you need to change the rb so that its type is image.

There is a lot more to learn about AWS3, and good tutorials are important.