Add new feature : Search
(1) The most important thing is to learn how to write a controller, and choose an efficient method.
(2) The search should be written in the header, in a separate file for easy reuse. Figure out what parameters to accept.
(3) After getting the parameters, configure the controller correctly, this step is very important.
(4) Remember to handle 0 search results.
A thought: this time, surprisingly, every time can be correctly rendered back to the product page, and did not specifically write render.

Translated with www.DeepL.com/Translator (free version)